\name{BicARE-package}
\alias{BicARE-package}
\alias{BicARE}
\docType{package}
\title{
BicARE
}
\description{
Biclustering Analysis and Results Exploration
}
\details{
\tabular{ll}{
Package: \tab BicARE\cr
Version: \tab 0.1.0\cr
Date: \tab 2008-06-05\cr
Depends: \tab R (>= 1.8.0), Biobase, multtest, GSEABase\cr
License: \tab GPL\cr
biocViews: \tab Microarray, Transcription, Statistics, Clustering\cr
URL: \tab http://bioinfo.curie.fr\cr
Packaged: \tab Tue Aug 19 14:21:38 2008; pgestraud\cr
Built: \tab R 2.8.0; i686-pc-linux-gnu; 2008-08-19 14:39:57; unix\cr
}

Index:
\preformatted{
FLOC                    Performs the FLOC algorithm
bicluster               Extract a bicluster
makeReport              Export the results as html files
residue                 Residue of a matrix
sample.bicData          Example data set for BicARE
sample.biclustering     Example biclustering object
testAnnot               Find samples annotations over-represented in
                        biclusters
testSet                 Find gene sets that are enriched in a bicluster
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{BicARE} \tab BicARE (source, pdf)\cr
}
}
\author{
Pierre Gestraud

Maintainer: Pierre Gestraud , <pierre.gestraud@curie.fr>
}
\keyword{ package }
