\name{sample.bicData}
\docType{data}
\alias{sample.bicData}

\title{Example data set for BicARE}

\description{A subset of sample.ExpressionSet from package Biobase. The data for 26 cases, labeled A to Z and 350 genes. Each case has
  three covariates: sex (male/female), type (case/control) and
  score (testing score).
}

\usage{sample.bicData}

\format{An ExpressionSet}

\keyword{datasets}