\name{sample.biclustering}
\docType{data}
\alias{sample.biclustering}

\title{Example biclustering object}

\description{A biclustering object created by the \code{\link{FLOC}}
  function on the sample.bicData with the following options
  : k=10, pGene = 0.3, pSample = 0.5, 
    r = 0.025, N = 8, M = 8, t = 1000.
}

\usage{sample.biclustering}

\format{a biclustering object}

\keyword{datasets}